chore: reorganized experimental components and create @ alias
