+++ package.json
@@ -17,7 +17,7 @@
         "build": "yarn clean && yarn build:ts && yarn build:babel",
         "build:link": "yarn build && yarn copy:package && yarn publish:yalc",
         "build:ts": "tsc --emitDeclarationOnly",
-        "build:babel": "babel src --out-dir dist --extensions \".ts,.tsx\"",
+        "build:babel": "babel src --out-dir dist --extensions \".ts,.tsx\" --ignore \"**/*.spec.tsx\" --ignore \"**/*.stories.tsx\"",
         "lint": "eslint -c .eslintrc.json 'src/**/*.{ts,tsx}'",
         "lint:fix": "eslint -c .eslintrc.json --fix 'src/**/*.{ts,tsx}'",
         "predeploy": "yarn build && yarn storybook:build",
@@ -82,6 +82,7 @@
         "babel-loader": "8.2.5",
         "babel-plugin-transform-class-properties": "6.24.1",
         "babel-plugin-transform-imports": "2.0.0",
+        "canvas": "2.11.2",
         "cypress": "12.0.2",
         "cypress-cucumber-preprocessor": "4.3.1",
         "cypress-plugin-tab": "1.0.5",
