+++ src/Experimental/Avatar/avatar.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { render, screen } from '@testing-library/react'
-import Avatar from './Avatar'
+import Avatar from '.'
 
 describe('Avatar', () => {
     it('should render first letter', () => {
