+++ src/Experimental/Avatar/index.tsx
@@ -1,3 +1,30 @@
-export { default as Avatar } from './Avatar'
+import { Avatar as FlipperAvatar } from '@/index'
+import { take } from 'ramda'
+import React, { CSSProperties, ReactNode } from 'react'
 
-// export more components, hooks or helpers below if needed
+export interface IProps {
+    /** The avatar name */
+    name?: string
+    /** A custom icon node */
+    icon?: ReactNode
+    /** An image source*/
+    src?: string
+    /** Custom styles */
+    style?: CSSProperties
+}
+
+const FIRST_LETTER = 1
+
+const Avatar = (props: IProps) => {
+    const getChild = () => {
+        if (props.name) {
+            return take(FIRST_LETTER, props.name.toUpperCase())
+        }
+
+        return props.icon
+    }
+
+    return <FlipperAvatar {...props}>{getChild()}</FlipperAvatar>
+}
+
+export default Avatar
