+++ src/Experimental/Button/button.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { render, screen } from '@testing-library/react'
-import Button from './Button'
+import Button from '.'
 
 describe('Button', () => {
     const NAME = 'btnName'
