+++ src/Experimental/Button/button.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from './Button'
+import Button from '.'
 
 export const button = () => {
     return <Button name='AddFoo' label='Add Foo' />
