+++ src/Experimental/Dialog/Dialog.tsx
@@ -1,8 +1,6 @@
-import DialogFlipper from '@core/Dialog'
+import { Dialog as DialogFlipper, Button, ThemeProvider } from '@/index'
 import React, { MouseEvent, ReactNode } from 'react'
 import { theme as Theme } from 'nginformatica-styleguide'
-import Button from '@core/Button'
-import ThemeProvider from '@core/ThemeProvider'
 
 const ACTION_REMOVE = Theme.colors.action.cancel
 const PRIMARY_DARK = Theme.colors.primary.main
