+++ src/Experimental/Dialog/index.tsx
@@ -1,5 +1,3 @@
 export { default as Dialog } from './Dialog'
 export { default as RemoveDialog } from './RemoveDialog'
 export { default as ConfirmDialog } from './ConfirmDialog'
-
-// export more components, hooks or helpers below if needed
