+++ src/Experimental/Fab/index.tsx
@@ -1,4 +1,2 @@
 export { default as Fab } from './Fab'
 export { default as FabWrapper } from './FabWrapper'
-
-// export more components, hooks or helpers below if needed
