+++ src/Experimental/Loading/Loading.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Loading from './Loading'
+import Loading from '.'
 
 export const defaultLoading = () => <Loading />
