--- src/Experimental/Sidebar/index.ts
@@ -1,1 +1,0 @@
-export { default as Sidebar } from './SidebarV2'
