+++ src/Experimental/index.ts
@@ -1,13 +1,13 @@
-export * from './Sidebar'
+export { default as Sidebar } from './Sidebar'
 export * from './Fab'
-export * from './Card'
-export * from './Button'
+export { default as Card } from './Card'
+export { default as Button } from './Button'
 export * from './Dialog'
-export * from './Avatar'
-export * from './Actions'
-export * from './Breadcrumbs'
-export * from './Loading'
-export * from './Panel'
-export * from './Skeleton'
-export * from './VisibilityIcon'
-export * from './NothingFound'
+export { default as Avatar } from './Avatar'
+export { default as Actions, Wrapper } from './Actions'
+export { default as Breadcrumbs } from './Breadcrumbs'
+export { default as Loading } from './Loading'
+export { default as Panel } from './Panel'
+export { default as Skeleton } from './Skeleton'
+export { default as VisibilityIcon } from './VisibilityIcon'
+export { default as NothingFound } from './NothingFound'
