+++ src/index.ts
@@ -63,3 +63,4 @@ export { default as DataTable } from './core/DataTable'
 export { default as StepCard } from './core/StepCard'
 export { default as StepCardSkeleton } from './core/StepCardSkeleton'
 export * as Experimental from './Experimental'
+export * as Icons from './icons'
