+++ tsconfig.json
@@ -26,9 +26,7 @@
         "skipLibCheck": true,
         "paths": {
             "@/*": ["./src/*"],
-            "@core/*": ["./src/core/*"],
-            "@experimental/*": ["./src/Experimental/*"],
         },
     },
-    "include": ["./src/*", "./src/core/**/*", "./src/lab/*", "./src/charts/*", ".src/*.d.ts", "./src/lib/**/*", "./src/Experimental/**/*.stories.tsx"],
+    "include": ["./src/*", "./src/core/**/*", "./src/lab/*", "./src/charts/*", ".src/*.d.ts", "./src/lib/**/*"],
 }
